feat(fetcher): Update logic for fetching and updating validator identifiers * Split main fetching module to submodules * Add usage of shared memory
